requirements.txt
azure-keyvault
azure-identity
tqdm
python-dotenv
text_unidecode
azure-ai-formrecognizer
langchain_openai
azure-ai-documentintelligence
langchain_community
azure-storage-blob
python-dotenv